+++ test/fixture/mock-generator/non-optional-nullable-object/builder.php
@@ -1,0 +1,9 @@
+<?php
+
+return new Eloquent\Phony\Mock\Builder\MockBuilder(
+    null,
+    array(
+        'methodA' => function (stdClass $first = null, $second) {},
+    ),
+    'Phony\Test\MockGeneratorNonOptionalNullableObject'
+);