+++ test/fixture/mock-generator/non-optional-nullable-object/expected.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Phony\Test;
+
+class MockGeneratorNonOptionalNullableObject
+implements \Eloquent\Phony\Mock\MockInterface
+{
+    public function methodA(
+        \stdClass $a0 = null,
+        $a1
+    ) {
+        $argumentCount = \func_num_args();
+        $arguments = array();
+
+        if ($argumentCount > 0) {
+            $arguments[] = $a0;
+        }
+        if ($argumentCount > 1) {
+            $arguments[] = $a1;
+        }
+
+        for ($i = 2; $i < $argumentCount; ++$i) {
+            $arguments[] = \func_get_arg($i);
+        }
+
+        return $this->_proxy->spy(__FUNCTION__)->invokeWith(
+            new \Eloquent\Phony\Call\Argument\Arguments($arguments)
+        );
+    }
+
+    private static $_uncallableMethods = array();
+    private static $_traitMethods = array();
+    private static $_customMethods = array();
+    private static $_staticProxy;
+    private $_proxy;
+}